backend-skill-gap/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
PyGithub==2.1.1
beautifulsoup4==4.12.2
lxml==4.9.3
reportlab==4.0.7
requests==2.31.0